feat: change SeekPayload Speed field type from float64 to float32 for consistency

+++ pkg/objects/video_channel.go
@@ -54,7 +54,7 @@ type VideoChannelObject interface {
 type SeekPayload struct {
 	PlaybackID string    `json:"playback_id"` //playback id to seek to
 	SeekTo     time.Time `json:"seek_to"`     //time to seek to
-	Speed      float64   `json:"speed"`       //speed to play the video, 1.0 is normal speed, 2.0 is double speed, 0.5 is half speed
+	Speed      float32   `json:"speed"`       //speed to play the video, 1.0 is normal speed, 2.0 is double speed, 0.5 is half speed
 	Reverse    bool      `json:"reverse"`     //if true, play the video in reverse
 }
